proc: add UnixSecond, UnixNanosecond fields
